Splitted HomeModule into seperate modules
